Feature: Option to choose stationary state if transition matrix is the identity (#8)
- Previously the stationary state was calculated from the transition
matrix, with an element of stochasticity – it is a random linear
combination of the Eigen vectors. Now if the transition matrix is not
set it is assumed to be the identity (no transitions) and instead the
ratio of the number of particles in each state can be set instead.
- Fix bad bug where the positions on the bezier path were not returned
correctly if it's length was less than 1. This caused particles to be
rendered motionless at the end of the previous segnebt and then jump to
the following segment.