FIx: Ground truth velocity projection (#9)
- The ground truth images were currently created using the simple path
position. However the kymograph sample path is different to the particle
true path. The kymograph sample path is by default only on one z plane,
this means the speeds of the particles recorded in the kymograph will be
slightly different to their true speed because they are travelling at an
angle to the plane.
- This PR adds a function to project the speeds to the xy plane and a
parameter in the ground truth generation functions to toggle this on or
off.
- Additionally, now because we don't need to worry about the ground
truth being incorrect the default particle path points generator has
been modified to include more extreme z variation.
- Additionally the number default path point generator has been modified
to modified the number of path points.