Chore(deps): Add missing deps; include group for cuda acceleration (#14)
- Add missing deps: `pyyaml`, `pillow`, `matplotlib`
- Add dependency group for accelerating microsim with gpu and driver
version 12
- Includes the `cupy` pre-release 14.0.0a1, this is due to earlier
versions of `cupy` not having the `cupy.asarray` function, which is
incompatible with `xarray` used in microsim.
- `microsim` is now coming from a branch on the fork
https://github.com/melisande-c/microsim/tree/fix/cupy to fix some issues
with cupy compatibility.

N.b.: For the `cupy` backend to work cuda toolkit has to already be
installed and sometimes has to be loaded by doing:
`module load cuda/12.4.1` (or whatever version is required).